@@ -1,23 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
+  <PropertyGroup>
+    <PreserveCompilationReferences>true</PreserveCompilationReferences>
+    <PreserveCompilationContext>true</PreserveCompilationContext>
+  </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="CSRedisCore" Version="3.3.0" />
-        <PackageReference Include="Dapper" Version="2.0.30" />
-        <PackageReference Include="HttpClientFactory" Version="1.0.1" />
+        <PackageReference Include="CSRedisCore" Version="3.6.5" />
+        <PackageReference Include="Dapper" Version="2.0.35" />
+        <PackageReference Include="HttpClientFactory" Version="1.0.3" />
         <PackageReference Include="Jint" Version="2.11.58" />
-        <PackageReference Include="LogDashboard" Version="1.4.3" />
-        <PackageReference Include="Microsoft.AspNetCore" Version="2.0.1" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.0.1" />
-        <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.0.1" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.0.0" />
-        <PackageReference Include="MySqlConnector" Version="0.61.0" />
-        <PackageReference Include="NLog.Extensions.Logging" Version="1.3.0" />
+        <PackageReference Include="LogDashboard" Version="1.4.5" />
+        <PackageReference Include="NLog.Web.AspNetCore" Version="4.9.3" />
+        <PackageReference Include="MySqlConnector" Version="1.0.1" />
         <PackageReference Include="RazorLight" Version="2.0.0-beta4" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.8" />
     </ItemGroup>
    
     
@@ -28,26 +29,8 @@
         <Content Update="nlog.config">
             <CopyToOutputDirectory>Always</CopyToOutputDirectory>
         </Content>
-        <Content Remove="backup\**" />
     </ItemGroup>
    
-    
-    <ItemGroup>
-      <_ContentIncludedByDefault Remove="webroot\appsettings.json" />
-    </ItemGroup>
+  
    
-    
-    <ItemGroup>
-      <Compile Remove="backup\**" />
-    </ItemGroup>
-   
-    
-    <ItemGroup>
-      <EmbeddedResource Remove="backup\**" />
-    </ItemGroup>
-   
-    
-    <ItemGroup>
-      <None Remove="backup\**" />
-    </ItemGroup>
 </Project>
