@@ -28,7 +28,7 @@ namespace JavaScriptViewEngine
     {
         private readonly RazorRenderEngine RazorRenderEngine;
 
-        public RazorEngineBuilder(Microsoft.AspNetCore.Hosting.IHostingEnvironment hostingEnvironment, IOptions<RazorEngineOption> options)
+        public RazorEngineBuilder(IWebHostEnvironment hostingEnvironment, IOptions<RazorEngineOption> options)
         {
             RazorRenderEngine = new RazorRenderEngine(hostingEnvironment);
         }
@@ -41,7 +41,7 @@ namespace JavaScriptViewEngine
 
     public class RazorRenderEngine : IRenderEngine
     {
-        private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IWebHostEnvironment _hostingEnvironment;
         private readonly RazorLightEngine _engine;
         private static readonly Logger logger;
 
@@ -71,7 +71,7 @@ namespace JavaScriptViewEngine
             logger = LogManager.GetCurrentClassLogger();
         }
 
-        public RazorRenderEngine(IHostingEnvironment hostingEnvironment)
+        public RazorRenderEngine(IWebHostEnvironment hostingEnvironment)
         {
             _hostingEnvironment = hostingEnvironment;
             _engine = new RazorLightEngineBuilder()
